feat: restructure into sub-commands
