+++ package.json
@@ -7,6 +7,7 @@
     "argparse": "^1.0.7",
     "base64url": "^2.0.0",
     "chalk": "^1.1.3",
+    "chloride": "^2.2.4",
     "commander": "^2.9.0",
     "ilp-plugin-ripple": "^2.1.0",
     "inquirer": "^1.1.3"
